+++ src/services/actions/index.ts
@@ -1,7 +1,7 @@
-import { TSettingsAction } from './reviever';
-import { TReviewerAction } from './settings';
+import { TReviewerAction } from "./reviever";
+import { TSettingsAction } from "./settings";
 
-export * from './reviever';
-export * from './settings';
+export * from "./reviever";
+export * from "./settings";
 
 export type TAllActions = TSettingsAction | TReviewerAction;
