+++ src/services/constants/index.ts
@@ -3,7 +3,7 @@ export const DELETE_FROM_BLACK_LIST = 'DELETE_FROM_BLACK_LIST';
 
 export const GET_REVIEWER__REQUEST = 'GET_REVIEWER__REQUEST';
 export const GET_REVIEWER__SUCCESS = 'GET_REVIEWER__SUCCESS';
-export const GET_REVIEWER__FAILURE = 'GET_REVIEWER__SUCCESS';
+export const GET_REVIEWER__FAILURE = 'GET_REVIEWER__FAILURE';
 
 export const UPDATE_FROM__LS = 'UPDATE_FROM__LS';
