+++ src/utils/constants.ts
@@ -1,2 +1,2 @@
-export const BASE_URL = 'https://api.github.com/repos/';
+export const BASE_URL = 'https://api.github.com/epos/';
 export const BASE_URL_END = '/contributors';
